+++ Scripts/Server/GetMatchupStrategy.cs
@@ -6,12 +6,12 @@ namespace Assets.Scripts.Server
     public class GetMatchupStrategy
     {
         public string Ships;
-        public int OpponentId;
+        public ulong OpponentId;
         public string[] Banned;
         public string Type;
         public long Hash;
 
-        public GetMatchupStrategy(string ships, int opponentId, string[] banned)
+        public GetMatchupStrategy(string ships, ulong opponentId, string[] banned)
         {
             this.Ships = ships;
             this.OpponentId = opponentId;
