+++ Scripts/Settings/StartingSettings.cs
@@ -20,7 +20,7 @@ namespace Assets.Scripts.Settings
         public List<ConfigData.ShipTypes> HumanShipTypes => HumanStartingShips.Keys.ToList(); // Barge, Carrier, Cruiser, Dreadnought, Drone, Factory, Fire Barge, Frigate, Gunship, Scout, Striker, Warp Gate
         public List<ConfigData.ShipTypes> BeeShipTypes => BeeStartingShips.Keys.ToList(); // Beehive, Bumblebee, Carpenter Bee, Honeybee, Hornet, Leafcutter, Queen, Wasp, Yellow Jacket
 
-        public StartingSettings(int userId) : base("starting-settings", userId)
+        public StartingSettings(ulong userId) : base("starting-settings", userId)
         {
         }
         protected override void ProcessData(string contents)
